requirements.txt
celery==5.3.6
channels==4.0.0
daphne==4.1.0
Django==4.2
django-admin-rangefilter==0.12.1
django-cors-headers==4.3.1
django-filter==24.1
django-phonenumber-field==7.3.0
djangorestframework==3.14.0
Markdown==3.5.2
phonenumbers==8.13.31
pillow==10.2.0
psycopg2-binary==2.9.9
PyJWT==2.8.0
redis==5.0.3
